use search params for size and colour
