+++ app/product/[productType]/page.tsx
@@ -66,15 +66,13 @@ export default async function ProductTypePage({
     );
 
     return (
-        <>
-            <ProductDetails
-                retrievedProduct={product}
-                initialSize={getInitialSize(displayName)}
-                initialColor={getInitialColour(displayName)}
-                variants={variants}
-                printifyImageId={imageId}
-            />
-        </>
+        <ProductDetails
+            retrievedProduct={product}
+            initialSize={getInitialSize(displayName)}
+            initialColor={getInitialColour(displayName)}
+            variants={variants}
+            printifyImageId={imageId}
+        />
     );
 }
